use static cast to avoid template mismatch on embedded devices
int isn't always int32_t on embedded devices, e.g. PlayDate.

+++ inkcpp/runner_impl.cpp
@@ -781,7 +781,7 @@ void runner_impl::step()
 				case Command::INT: {
 					int val = read<int>();
 					if (_evaluation_mode) {
-						_eval.push(value{}.set<value_type::int32>(val));
+						_eval.push(value{}.set<value_type::int32>(static_cast<int32_t>(val)));
 					}
 					// TEST-CASE B006 don't print integers
 				} break;
@@ -1011,7 +1011,7 @@ void runner_impl::step()
 					hash_t functionName = read<hash_t>();
 
 					// Interpret flag as argument count
-					int numArguments = ( int ) flag;
+					int numArguments = static_cast<int>(flag);
 
 					// find and execute. will automatically push a valid if applicable
 					auto* fn = _functions.find(functionName);
@@ -1200,11 +1200,11 @@ void runner_impl::step()
 					// Push the visit count for the current container to the top
 					//  is 0-indexed for some reason. idk why but this is what ink expects
 					_eval.push(
-					    value{}.set<value_type::int32>(( int ) _globals->visits(_container.top().id) - 1)
+					    value{}.set<value_type::int32>(static_cast<int32_t>(_globals->visits(_container.top().id) - 1))
 					);
 				} break;
 				case Command::TURN: {
-					_eval.push(value{}.set<value_type::int32>(( int ) _globals->turns()));
+					_eval.push(value{}.set<value_type::int32>(static_cast<int32_t>(_globals->turns())));
 				} break;
 				case Command::SEQUENCE: {
 					// TODO: The C# ink runtime does a bunch of fancy logic
@@ -1229,7 +1229,7 @@ void runner_impl::step()
 					container_t container = read<container_t>();
 
 					// Push the read count for the requested container index
-					_eval.push(value{}.set<value_type::int32>(( int ) _globals->visits(container)));
+					_eval.push(value{}.set<value_type::int32>(static_cast<int32_t>(_globals->visits(container))));
 				} break;
 				case Command::TAG: {
 					_tags.push() = read<const char*>();
